+++ spec/unit/alf-core/test_aggregator.rb
@@ -10,7 +10,8 @@ module Alf
     ]}
 
     it "should keep track of registered aggregators" do
-      Aggregator.aggregators.should_not be_empty
+      Aggregator.all.should be_a(Array)
+      Aggregator.all.should_not be_empty
       Aggregator.each do |agg|
         agg.should be_a(Class)
       end
@@ -21,31 +22,31 @@ module Alf
     end
 
     describe "coerce" do
-     
+
       subject{ Aggregator.coerce(arg) }
-      
+
       describe "from an Aggregator" do
         let(:arg){ Aggregator.sum{a} }
         it{ should eq(arg) }
       end
-      
+
       describe "from a String" do
         let(:arg){ "sum{a}" }
 
-        it { 
-          should be_a(Aggregator::Sum) 
+        it {
+          should be_a(Aggregator::Sum)
         }
 
         it 'should have source code' do
           subject.has_source_code!.should eq("sum{a}")
         end
 
-        specify{ 
-          subject.aggregate(input).should eql(7) 
+        specify{
+          subject.aggregate(input).should eql(7)
         }
 
       end
-      
+
     end
 
     it 'should implement optimistic equality based on source code' do